refactor: move query datetime to URL objects
